Move ResourceUtilizationInstruments to Shared project
