+++ src/Shared/Instruments/README.md
@@ -1,0 +1,11 @@
+# Diagnostic IDs
+
+Defines various diagnostic IDs reported by this repo.
+
+To use this in your project, add the following to your `.csproj` file:
+
+```xml
+<PropertyGroup>
+  <InjectSharedDiagnosticIds>true</InjectSharedInstruments>
+</PropertyGroup>
+```
