+++ src/Shared/Instruments/ResourceUtilizationInstruments.cs
@@ -1,9 +1,7 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
-
-namespace Microsoft.Extensions.Diagnostics.ResourceMonitoring;
+namespace Microsoft.Shared.Instruments;
 
 /// <summary>
 /// Represents the names of instruments published by this package.
